refactor(types): 整合 AI 类型定义到 shared 模块，删除重复和僵尸代码
